+++ docs/index.html
@@ -4,5 +4,5 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   </head>
   <body>
-  <script type="text/javascript" src="vendor.bundle.da3c15f7.js"></script><script type="text/javascript" src="app.bundle.e99c615d.js"></script></body>
+  <script type="text/javascript" src="vendor.bundle.a21f16f8.js"></script><script type="text/javascript" src="app.bundle.f43781c8.js"></script></body>
 </html>
